refactor(proto): restore common-type conversions as From/TryFrom in proto-models
`UnnestOptions`, `TableReference`, `StringifiedPlan`, `JoinType`,
`JoinConstraint`, `NullEquality`, `CsvOptions`, `JsonOptions` and the parquet
options types all live in `datafusion-common`, which sits below
`datafusion-proto-models` in the crate graph and so cannot host the impls.
They move onto the local proto type in `datafusion-proto-models` instead — the
arrangement `datafusion-proto-common` already uses for `ScalarValue` and
`Statistics` — and go back to being plain `From` / `TryFrom`, the shape they
had in 54.1.0.

This is also the only crate that can express
`TryFrom<&[PartitionedFile]> for protobuf::FileGroup`. `datafusion-datasource`
cannot: `&T` is `#[fundamental]` but `[T]` is not, so `&[PartitionedFile]`
counts as foreign there. Here the *self* type is local, which is all the orphan
rule needs, and staying generic over the element (`&T: TryInto<protobuf::PartitionedFile>`)
means this crate never has to name `PartitionedFile`, which sits above it.
Callers get the 54.1.0 spelling back verbatim.

New `datafusion_proto_models::{from_proto, to_proto}` modules; the crate gains
a direct `datafusion-common` dependency (already present transitively via
`datafusion-proto-common`).

Part of #24019.
